+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * Try `nix-tree nixpkgs#hello'
 * change: `nix-tree` now requires an explicit `--derivation` flag to work on store derivations rather than their outputs.
   * This is to be more consistent `nix path-info`.
+* fix: Fix timeout when running external command for the 'yank' functionality (issue: [#25][])
 
 [#27]: https://github.com/utdemir/nix-tree/issues/27